Reorganized File Structure and Moved Harmony Patches to Class based Dedicated Files
