+++ BabyStepsMultiplayerClient/Patches/Patch_SaveGod.cs
@@ -1,0 +1,21 @@
+﻿using HarmonyLib;
+using Il2Cpp;
+
+namespace BabyStepsMultiplayerClient.Patches
+{
+    [HarmonyPatch]
+    internal class Patch_SaveGod
+    {
+        [HarmonyPrefix]
+        [HarmonyPatch(typeof(SaveGod), nameof(SaveGod.LoadSave))]
+        private static bool LoadSave_Prefix(SaveGod __instance)
+        {
+            Core _core = Core.thisInstance;
+            if (_core.client != null)
+                _core.Disconnect();
+
+            // Run Original
+            return true;
+        }
+    }
+}
